requirements.txt
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.0
cycler==0.12.1
fonttools==4.55.3
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gspread==6.1.4
httplib2==0.22.0
idna==3.10
importlib_resources==6.5.2
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.4
numpy==2.0.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-pptx==1.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
XlsxWriter==3.2.0
zipp==3.21.0
